[Silver III] Title: 주유소, Time: 36 ms, Memory: 2552 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.c
@@ -3,13 +3,13 @@
 int main() {
     int N;
     scanf("%d", &N);
-    int dist[N], cost[N], min, ans = 0;
-    for (int i = 1; i < N; i++) scanf("%d", &dist[i]);
-    for (int i = 0; i < N; i++) scanf("%d", &cost[i]);
+    long dist[N], cost[N], min, ans = 0;
+    for (int i = 1; i < N; i++) scanf("%ld", &dist[i]);
+    for (int i = 0; i < N; i++) scanf("%ld", &cost[i]);
     min = cost[0];
     for (int i = 1; i < N; i++) {
         ans += dist[i] * min;
         min = min > cost[i] ? cost[i] : min;
     }
-    printf("%d", ans);
+    printf("%ld", ans);
 }
